I've ever-so slightly modified how names are aligned on the contact list: - The bottoms of letters (gjpq...) no longer get cut off with very small fonts (9pt) - Average fonts should appear more centered when the grid is turnd on (10-11pt)
Certain fonts still appear too low, but I am unsure how to solve this.  The majority of fonts
should work great at this time though.

Status circles are now displaying idle time.  I've tried very hard to get the idle time looking
good at the most used font sizes (9-12), but it has proven to be very challenging.  I think the
results are pretty good though, and better looking than 1.x.

I've switched the status circle from times to Lucide Grande.  Lucide Grande is unbelievably
crisp at small font sizes - resulting in perfecty scaled status circles when the contact list is
using a 9 point font!

The amount of indent made available for left views (such as status circles) is now scaled to
match the row height.  This prevents users with small fonts from suffering with an abnormally
large looking indent.

Idle times are now updated correctly by the TOC account code.

+++ adium/Source/AIAdium.h
@@ -127,7 +127,7 @@ typedef enum {
 
 @protocol AIHandleLeftView //Draws to the left of a handle
     - (void)drawInRect:(NSRect)inRect;
-    - (int)widthForHeight:(int)inHeight;
+    - (float)widthForHeight:(int)inHeight;
 @end
 
 @protocol AIHandleRightView //Draws to the right of a handle
